requirements.txt
fastapi
uvicorn
python-multipart
Pillow
transformers
torch
sentencepiece 